Shiny_App prediction model performance
In TikTok dataset, there are different variables regarding to trending TikTok tracks. My research question was to understand the predictors that predict the popularity of a track. In order to understand which variables (and the interaction of which variables) create the best model, I use this app. Depending on whther you aim to get a first order model or second order model, you are able to see the best combinations on a plot. The predictors with lower RMSE, are more appropriate to be used in our model.